Centralize MCP path boundary checks (#3753)
