+++ changelog.d/unreleased/3753.fixed.md
@@ -1,0 +1,17 @@
+---
+category: fixed
+issues:
+  - 3753
+affected:
+  - src/CodeIndex/Mcp/McpPathBoundary.cs
+  - src/CodeIndex/Mcp/McpToolHandlers.cs
+  - tests/CodeIndex.Tests/McpServerTests.cs
+---
+
+## English
+
+- **MCP filesystem path containment now routes through one boundary helper (#3753)** — required workspace-relative paths, client-root file URIs, and `index` containment checks now share symlink-aware normalization before deciding whether a path remains inside the allowed directory.
+
+## 日本語
+
+- **MCP filesystem path の containment が単一の boundary helper を通るようになりました (#3753)** — 必須の workspace-relative path、client root の file URI、`index` の containment 判定は、許可ディレクトリ内に留まるか判断する前に symlink-aware な正規化を共有します。
